requirements.txt
django
djangorestframework
drf-yasg
celery
redis
pylint
pylint-django
google-auth